Update api-wasm.md
Trying to run the example generated in an error: 
SyntaxError: The requested module 'in3-wasm' is expected to be of type CommonJS,
which does not support named exports. 
CommonJS modules can be imported by importing the default export.
For example: import pkg from 'in3-wasm'; const { IN3 } = pkg;

Thus, added this change, this is executing as expected.